Documents/Projekty/Laura/requirements.txt
langchain
langchain-openai
langgraph
python-dotenv
fastapi
uvicorn
pydantic
